Create Password endpoint
